+++ java/src/test/java/io/github/cboudereau/dataseries/snippets/IntersectionTest.java
@@ -9,28 +9,13 @@ import org.junit.jupiter.api.Test;
 import io.github.cboudereau.dataseries.Series;
 import io.github.cboudereau.dataseries.UnionResult;
 
-public class SamplesTest {
-    @Test
-    public void simple() {
-        var s1 = List.of(Series.datapoint(3, 50));
-        var s2 = List.of(Series.datapoint(4, 100), Series.datapoint(7, 110));
-
-        var actual = Series.union(s1, s2, x -> x).stream().toArray();
-
-        var expected = List.of(
-                Series.datapoint(3, UnionResult.leftOnly(50)),
-                Series.datapoint(4, UnionResult.both(50, 100)),
-                Series.datapoint(7, UnionResult.both(50, 110))).toArray();
-
-        assertArrayEquals(expected, actual);
-    }
-
+public class IntersectionTest {
     @Test
     public void intersection() {
         var s1 = List.of(Series.datapoint(3, 50));
         var s2 = List.of(Series.datapoint(4, 100), Series.datapoint(7, 110));
 
-        var actual = Series.union(s1, s2, SamplesTest::toTuple).stream().filter(x -> x.data().isPresent())
+        var actual = Series.union(s1, s2, IntersectionTest::toTuple).stream().filter(x -> x.data().isPresent())
                 .map(x -> Series.datapoint(x.point(), x.data().get())).toArray();
 
         var expected = List.of(
